Make recordpeeker work on Windows.
There were two problems.  The first was that it was trying to
|call| mitmdump, a script which is installed by default at
$PYTHONHOME/Scripts/mitmdump.  This is not part of %PATH% on
Windows, however, so it can't be found.  So this patch adds
mitmdump_shim.py which is the same as the mitmdump script, and
we use subprocess.call on that instead.

The second problem this fixes is that python scripts are not
executable by default on Windows, so you can't just call the
script.  You need to use sys.executable.